requirements.txt
fastapi
uvicorn[standard]
pydantic
torch
transformers
sentence-transformers
faiss-cpu
scikit-learn
datasets